+++ lib/property.php
@@ -675,7 +675,7 @@ class Property implements \D2U_Helper\ITranslationHelper
                         .'ON main.property_id = default_lang.property_id AND default_lang.clang_id = '. rex_config::get('d2u_helper', 'default_lang') .' '
                     .'WHERE target_lang.property_id IS NULL '
                     .'ORDER BY default_lang.name';
-            $clang_id = rex_config::get('d2u_helper', 'default_lang');
+            $clang_id = (int) rex_config::get('d2u_helper', 'default_lang');
         }
         $result = rex_sql::factory();
         $result->setQuery($query);
@@ -701,7 +701,7 @@ class Property implements \D2U_Helper\ITranslationHelper
 
             $parameterArray = [];
             $parameterArray['property_id'] = $this->property_id;
-            $this->url = rex_getUrl($d2u_immo->getConfig('article_id'), $this->clang_id, $parameterArray, '&');
+            $this->url = rex_getUrl((int) $d2u_immo->getConfig('article_id'), $this->clang_id, $parameterArray, '&');
         }
 
         if ($including_domain) {
